+++ examples/memory-store/src/store.rs
@@ -1,4 +1,4 @@
-#[cfg(any(feature = "email", feature = "otp", feature = "sms"))]
+#[cfg(any(feature = "email", feature = "sms"))]
 use crate::models::MyEmailChallenge;
 #[cfg(any(
     feature = "email",
@@ -33,7 +33,7 @@ use tokio::sync::RwLock;
 pub struct MemoryStore {
     sessions: Arc<RwLock<HashMap<Uuid, MyLoginSession>>>,
     users: Arc<RwLock<HashMap<Uuid, MyUser>>>,
-    #[cfg(any(feature = "email", feature = "otp", feature = "sms"))]
+    #[cfg(any(feature = "email", feature = "sms"))]
     challenges: Arc<RwLock<HashMap<String, MyEmailChallenge>>>,
     #[cfg(feature = "oauth")]
     oauth_tokens: Arc<RwLock<HashMap<Uuid, MyOAuthToken>>>,
@@ -75,12 +75,12 @@ impl IntoResponse for MemoryStoreError {
 
 impl AutheryStore for MemoryStore {
     type User = MyUser;
-    #[cfg(any(feature = "email", feature = "otp"))]
+    #[cfg(feature = "email")]
     type UserEmail = MyUserEmail;
     #[cfg(feature = "sms")]
     type UserPhone = MyUserPhone;
     type LoginSession = MyLoginSession;
-    #[cfg(any(feature = "email", feature = "otp", feature = "sms"))]
+    #[cfg(any(feature = "email", feature = "sms"))]
     type EmailChallenge = MyEmailChallenge;
     #[cfg(feature = "oauth")]
     type OAuthToken = MyOAuthToken;
@@ -253,7 +253,7 @@ impl AutheryStore for MemoryStore {
         Ok(())
     }
 
-    #[cfg(any(feature = "email", feature = "otp", feature = "sms"))]
+    #[cfg(any(feature = "email", feature = "sms"))]
     async fn create_challenge(
         &self,
 
@@ -275,7 +275,7 @@ impl AutheryStore for MemoryStore {
         Ok(challenge)
     }
 
-    #[cfg(any(feature = "email", feature = "otp", feature = "sms"))]
+    #[cfg(any(feature = "email", feature = "sms"))]
     async fn consume_challenge(
         &self,
         code: String,
@@ -288,7 +288,7 @@ impl AutheryStore for MemoryStore {
         Ok(challenge)
     }
 
-    #[cfg(any(feature = "email", feature = "otp"))]
+    #[cfg(feature = "email")]
     async fn get_user_emails(&self, user_id: &Uuid) -> Result<Vec<MyUserEmail>, Self::Error> {
         let users = self.users.read().await;
 
@@ -579,7 +579,7 @@ impl AutheryStore for MemoryStore {
     }
 
     // user store
-    #[cfg(any(feature = "email", feature = "otp"))]
+    #[cfg(feature = "email")]
     async fn get_user_by_email_address(
         &self,
         address: &str,
@@ -594,7 +594,7 @@ impl AutheryStore for MemoryStore {
         }))
     }
 
-    #[cfg(any(feature = "email", feature = "otp"))]
+    #[cfg(feature = "email")]
     async fn create_user_by_email_address(
         &self,
         address: &str,